+++ libtitania-x3d/Titania/X3D/Components/Texturing/MultiTextureCoordinate.h
@@ -115,7 +115,7 @@ public:
 
 	virtual
 	void
-	enable (size_t, size_t, const TexCoordArray &) const final
+	enable (int32_t, size_t, const TexCoordArray &) const final
 	{ }
 
 	virtual
@@ -124,7 +124,7 @@ public:
 
 	virtual
 	void
-	disable (size_t) const final
+	disable (int32_t) const final
 	{ }
