+++ libtitania-x3d/Titania/X3D/Components/Texturing/X3DTextureTransformNode.h
@@ -69,7 +69,7 @@ public:
 
 	virtual
 	void
-	draw (size_t);
+	draw (int32_t);
 
 
 protected:
